We can now change the alarm but the modal dialog doesn't show the correct sound selected and neither does the table view. It doesn't persist either.

+++ Classes/PieChartLayer.m
@@ -9,7 +9,9 @@
 #import "PieChartLayer.h"
 
 
-static inline double radians (double degrees) {return degrees * M_PI/180;}
+static inline double radians(double degrees) {
+    return degrees * M_PI / 180;
+}
 
 
 @implementation PieChartLayer
@@ -20,51 +22,51 @@ static inline double radians (double degrees) {return degrees * M_PI/180;}
 #pragma mark -
 #pragma mark CALayer
 
-- (id)initWithLayer:(id)layer {
-	if((self = [super initWithLayer:layer])) {
-		if([layer isKindOfClass:[PieChartLayer class]]) {
-			PieChartLayer *other = (PieChartLayer*)layer;
-			self.startAngle = other.startAngle;
-			self.endAngle = other.endAngle;
-		}
-	}
-	return self;
+-(id)initWithLayer:(id)layer {
+    if ((self = [super initWithLayer:layer])) {
+        if ([layer isKindOfClass:[PieChartLayer class]]) {
+            PieChartLayer *other = (PieChartLayer *) layer;
+            self.startAngle = other.startAngle;
+            self.endAngle   = other.endAngle;
+        }
+    }
+    return self;
 }
 
-+ (BOOL)needsDisplayForKey:(NSString *)key {
-	if ([key isEqualToString:@"startAngle"]
-		|| [key isEqualToString:@"endAngle"]) {
++(BOOL)needsDisplayForKey:(NSString *)key {
+    if ([key isEqualToString:@"startAngle"]
+        || [key isEqualToString:@"endAngle"]) {
         return YES;
     }
-	else {
+    else {
         return [super needsDisplayForKey:key];
     }
 }
 
-- (void)drawInContext:(CGContextRef)ctx {
-	
-	/* Getting some values */
-	CGFloat radius = fmin(self.bounds.size.width, self.bounds.size.height)/4;
-	CGPoint centerPoint = CGPointMake(CGRectGetWidth(self.bounds)/2, CGRectGetHeight(self.bounds)/3);	
-	CGFloat sAngle = radians(self.startAngle);
-	CGFloat eAngle = radians(self.endAngle);
-	
-	/* Drawing arc */
-	CGContextMoveToPoint(ctx, centerPoint.x, centerPoint.y);
-	CGContextAddArc(ctx, centerPoint.x, centerPoint.y, radius, sAngle, eAngle, 1);
-	CGContextClosePath(ctx);
-	
-	/* Filling it */
-	CGContextSetFillColorWithColor(ctx, [[[[UIColor alloc] initWithRed:0 green:0 blue:0 alpha:0.3] autorelease] CGColor]);
-	CGContextFillPath(ctx);	
+-(void)drawInContext:(CGContextRef)ctx {
+
+    /* Getting some values */
+    CGFloat radius      = (CGFloat) (fmin(self.bounds.size.width, self.bounds.size.height) / 4);
+    CGPoint centerPoint = CGPointMake(CGRectGetWidth(self.bounds) / 2, CGRectGetHeight(self.bounds) / 3);
+    CGFloat sAngle      = (CGFloat) radians(self.startAngle);
+    CGFloat eAngle      = (CGFloat) radians(self.endAngle);
+
+    /* Drawing arc */
+    CGContextMoveToPoint(ctx, centerPoint.x, centerPoint.y);
+    CGContextAddArc(ctx, centerPoint.x, centerPoint.y, radius, sAngle, eAngle, 1);
+    CGContextClosePath(ctx);
+
+    /* Filling it */
+    CGContextSetFillColorWithColor(ctx, [[[[UIColor alloc] initWithRed:0 green:0 blue:0 alpha:0.3] autorelease] CGColor]);
+    CGContextFillPath(ctx);
 }
 
 #pragma mark -
 #pragma mark Public API
 
-- (id) lastValueForKey:(NSString*)key {
-	PieChartLayer *last = (PieChartLayer*)self.presentationLayer;
-	return [last valueForKey:key];
+-(id)lastValueForKey:(NSString *)key {
+    PieChartLayer *last = (PieChartLayer *) self.presentationLayer;
+    return [last valueForKey:key];
 }
 
 @end
